% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_CombineAssignOutput.R
\name{combine_assign_output}
\alias{combine_assign_output}
\title{Combine Taxa Assignment for Same ASV Using Different Databases}
\usage{
combine_assign_output(fileLoc = NULL, numCores = 1)
}
\arguments{
\item{fileLoc}{The location of a file in a directory where all of the 'taxaAssign'
files are located.}

\item{numCores}{The number of cores used to run the function (default = 1,
Windows systems can only use a single core).}
}
\value{
This function produces a '2023_08_03_0913_taxaAssignCombined.tsv' and
a '2023_08_03_0913_taxaAssignCombined.txt' file in the selected target
directory.
}
\description{
This function takes a file selection and then uses all 'taxaAssign' files in
that directory and combines them into a single output
'taxaAssignCombined.tsv' file.
}
\details{
The User Input: This function requires a file in a directory where all 'taxaAssign'
files in that directory will be combined.
}
\note{
When running DBTCShiny functions the paths for the files selected cannot have
whitespace! File folder locations should be as short as possible (close to
the root directory) as some functions do not process long naming conventions.
Also, special characters should be avoided (including question mark, number
sign, exclamation mark). It is recommended that dashes be used for
separations in naming conventions while retaining underscores for use as
information delimiters (this is how DBTC functions use underscore). There
are several key character strings used in the DBTC pipeline, the presence of
these strings in file or folder names will cause errors when running DBTC
functions.

The following strings are those used in DBTC and should not be used in file
or folder naming:
- _BLAST
- _taxaAssign
- _taxaCombined
- _taxaReduced
}
\examples{
\dontrun{
combine_assign_output()
combine_assign_output(fileLoc = NULL,   numCores = 1)
}

}
\references{
<https://github.com/rgyoung6/DBTC>
Young, R. G., Hanner, R. H. (Submitted October 2023). Title Here. Biodiversity Data Journal.
}
\seealso{
dada_implement()
combine_dada_output()
make_BLAST_DB()
seq_BLAST()
taxon_assign()
reduce_taxa()
combine_reduced_output()
}
\author{
Robert G. Young
}
