% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_launchDBTC.R
\name{launchDBTCShiny}
\alias{launchDBTCShiny}
\title{Launch DBTCShiny}
\usage{
launchDBTCShiny(verbose = TRUE)
}
\arguments{
\item{verbose}{If set to TRUE then there will be output to the R console, if
FALSE then this reporting data is suppressed (Default TRUE).}
}
\value{
There are no values or files returned from this function
}
\description{
This function launches the DBTCShiny Application
}
\details{
This function launches a DBTCShiny Application which allows a user to run the 'DBTC'
functions and process high throughput sequencing data.
}
\note{
This is a wrapper function which launches the DBTCShiny package as a 'Shiny'
application in the systems default browser program
}
\examples{
if(interactive()){
launchDBTCShiny()
}

}
\references{
<https://github.com/rgyoung6/DBTC>
Young, R. G., Hanner, R. H. (Submitted October 2023). Metabarcoding analysis
using Dada-BLAST-Taxon Assign-Condense shiny Application (DBTCShiny). Biodiversity Data Journal.
}
\seealso{
dada_implement()
combine_dada_output()
make_BLAST_DB()
seq_BLAST()
taxon_assign()
combine_assign_output()
reduce_taxa()
combine_reduced_output()
}
\author{
Robert G. Young
}
