% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_make_BLAST_DB.R
\name{make_BLAST_DB}
\alias{make_BLAST_DB}
\title{Make a BLAST Database}
\usage{
make_BLAST_DB(
  fileLoc = NULL,
  makeblastdbPath = "makeblastdb",
  taxaDBLoc = NULL,
  inputFormat = NULL,
  dbName = NULL,
  minLen = 100
)
}
\arguments{
\item{fileLoc}{The location of a file in a directory where all fasta files
will be used to construct a BLASTable database. Default = NULL}

\item{makeblastdbPath}{The local path for the blast+ makeblastdbPath program}

\item{taxaDBLoc}{The location of the NCBI taxonomic data base (accessionTaxa.sql
see the main DBTCShiny page for details).}

\item{inputFormat}{This will either be NCBI formatted fasta file (header
example >uniqueID|NCBItaxaID|Genus|species) or a MACER
(header example >uniqueID|other_ID|Genus|species|Other_info|markerOrDatabase)
formatted file.}

\item{dbName}{A short 6-8 alpha character name used when building a database.}

\item{minLen}{The minimum sequence length used to construct the BLAST database.}
}
\value{
The output from this function includes a folder with the BLAST database named
according to the submitted dbName
}
\description{
This function takes a fasta file (in MACER format or NCBI format) and
establishes a database upon which a BLAST search can be completed.
}
\details{
The user inputs the location of a file in a directory that contains properly formatted
fasta files which can be used to construct a dataset that can BLASTED. The
NCBI blast+ program, makeblastdb and the NCBI taxonomic database are required to
run this script (see the main page of the DBTCShiny for instructions on how to
obtain these resources).
}
\note{
When running DBTCShiny functions the paths for the files selected cannot have
whitespace! File folder locations should be as short as possible (close to
the root directory) as some functions do not process long naming conventions.
Also, special characters should be avoided (including question mark, number
sign, exclamation mark). It is recommended that dashes be used for
separations in naming conventions while retaining underscores for use as
information delimiters (this is how DBTC functions use underscore). There
are several key character strings used in the DBTC pipeline, the presence of
these strings in file or folder names will cause errors when running DBTC
functions.

The following strings are those used in DBTC and should not be used in file
or folder naming:
- _BLAST
- _taxaAssign
- _taxaCombined
- _taxaReduced
}
\examples{
\dontrun{
make_BLAST_DB()
make_BLAST_DB(fileLoc = NULL, makeblastdbPath = "makeblastdb", taxaDBLoc = NULL,
inputFormat = NULL, dbName = NULL, minLen = 100)
}

}
\references{
<https://github.com/rgyoung6/DBTC>
Young, R. G., Hanner, R. H. (Submitted October 2023). Title Here. Biodiversity Data Journal.
}
\seealso{
dada_implement()
combine_dada_output()
seq_BLAST()
taxon_assign()
combine_assign_output()
reduce_taxa()
combine_reduced_output()
}
\author{
Robert G. Young
}
