% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_ReduceTaxa.R
\name{reduce_taxa}
\alias{reduce_taxa}
\title{Reduce Taxa Assignment}
\usage{
reduce_taxa(fileLoc = NULL, numCores = 1)
}
\arguments{
\item{fileLoc}{The location of a file in a directory where all of the 'taxaAssign'
and/or 'taxaAssignCombine' files are located.}

\item{numCores}{The number of cores used to run the function (default = 1,
Windows systems can only use a single core).}
}
\value{
This function produces a 'taxa_reduced' file for every 'taxaAssign' or
'taxaAssignCombine' present in the target directory.
}
\description{
This function takes a file selection and then uses all 'taxaAssign' and/or
'taxaAssignCombine' files in that directory and reduces all ASV with the same
taxonomic assignment into a single result and places these results in a
'taxaReduced' file for each of the target files in the directory.
}
\details{
This function requires a file in a directory where all 'taxaAssign'
and/or 'taxaAssignCombine' files in that directory will be combined. All records
with the same taxonomic result will be combined. The BLAST values in parentheses
("Num_Rec", "Coverage", "Identity", "Max_eVal") are combine by the mean number of records,
the minimum coverage and identity, and the maximum eValue.
}
\note{
When running DBTCShiny functions the paths for the files selected cannot have
whitespace! File folder locations should be as short as possible (close to
the root directory) as some functions do not process long naming conventions.
Also, special characters should be avoided (including question mark, number
sign, exclamation mark). It is recommended that dashes be used for
separations in naming conventions while retaining underscores for use as
information delimiters (this is how DBTC functions use underscore). There
are several key character strings used in the DBTC pipeline, the presence of
these strings in file or folder names will cause errors when running DBTC
functions.

The following strings are those used in DBTC and should not be used in file
or folder naming:
- _BLAST
- _taxaAssign
- _taxaCombined
- _taxaReduced
}
\examples{
\dontrun{
reduce_taxa()
reduce_taxa(fileLoc = NULL,   numCores = 1)
}

}
\references{
<https://github.com/rgyoung6/DBTC>
Young, R. G., Hanner, R. H. (Submitted October 2023). Title Here. Biodiversity Data Journal.
}
\seealso{
dada_implement()
combine_dada_output()
make_BLAST_DB()
seq_BLAST()
taxon_assign()
combine_assign_output()
combine_reduced_output()
}
\author{
Robert G. Young
}
